requirements.txt
spacy==3.1.2
# en_core_web_sm==3.0.0
squarify==0.4.3
emoji==1.2.0
nltk==3.5
praw==7.4.0
pandas==1.0.5
matplotlib==3.2.2
data==0.4
pyyaml==5.4.1
marshmallow==3.13.0 # for serialize and deserialize json to python data objects
marshmallow-dataclass==8.5.3